+++ ShutTheDuckUpBreakOut/Assets/Script/Items/Objects_Guns.cs
@@ -14,7 +14,7 @@ public class Objects_Guns : ScriptableObject
     public int SpeedBullet;  // the damage you do
     public int ShotsInMagasin; // how many hits you can hit
     public int bulletFiredPerShot; // how many shots per Bullet
-    public int knockBack; // knockback
+    public float knockBack; // knockback
     public float bulletSPread; // bulletSpread
     public bool HoldToFire;
     public Vector3 BulletSize;
